Update predict.py
bridging test
