+++ model_bridging/model_bridging/helpers.py
@@ -1,42 +1,7 @@
 import numpy as np
-import pandas as pd
-import boto3
-import joblib
-import tempfile
+import pandas as pds
 from typing import List
-
-
-def get_bucket_and_key_from_s3_uri(uri):
-    bucket, key = uri.split("/", 2)[-1].split("/", 1)
-    return bucket, key
-
-
-def tokenize_pd_code(df):
-
-    df[['pd_code_1', 'pd_code_2']] = pd.DataFrame(
-        df['PrimaryDiagnosisCode'].fillna("na.na").str.split(".", expand=True))
-    return df.drop('PrimaryDiagnosisCode', axis=1)
-
-
-def get_date_diff(col1, col2, interval):
-    """difference between dates specified by the interval in ['Y','D','M']
-    col1 and col2 are date colummns and col2 > col1"""
-
-    return (col2 - col1) / np.timedelta64(1, interval)
-
-
-def add_emp_tenure_to_df(df):
-    """
-    returns a df with employment tenure column appeneded
-    """
-    df["emp_tenure"] = get_date_diff(
-        df["InsuredHireDate"], df["LossDate"], interval="D"
-    )
-
-    return df.drop("InsuredHireDate", axis=1)
-
-
-# joblib needs this class
+#joblib needs this class
 class CategoricalGrouper:
 
     """we are fitting the categorical grouper on a reference data
@@ -46,7 +11,7 @@ class CategoricalGrouper:
     def __init__(self, min_freq):
         self.min_freq = min_freq
         self.is_fitted = False
-        self._non_oov_dict = None
+
 
     def fit(self, X, categorical: List):
         _non_oov_dict = dict()
@@ -55,9 +20,11 @@ class CategoricalGrouper:
             _df.columns = ["index", "counts"]
             non_oov = list(_df.loc[_df['counts'] >= self.min_freq, "index"])
             _non_oov_dict[col] = non_oov
+#             X.loc[~X[col].isin(non_oov), col] = 'OOV'
         self._non_oov_dict = _non_oov_dict
         self.is_fitted = True
         return None
+
 
     def transform(self, X, categorical: List):
         if self.is_fitted:
@@ -71,10 +38,3 @@ class CategoricalGrouper:
                 "The categorical grouper needs to be fit before calling .transform()"
             )
 
-
-def download_model_from_s3(bucket_name, key):
-    bucket = boto3.resource("s3").Bucket(bucket_name)
-    with tempfile.NamedTemporaryFile() as fp:
-        bucket.download_fileobj(key, fp)
-        loaded_model = joblib.load(fp.name)
-    return loaded_model
