+++ model_bridging/model_bridging/predict.py
@@ -1,21 +1,52 @@
 def predict(**kwargs):
     import pandas as pd
     import numpy as np
+    import boto3
+    import joblib
+    import tempfile
+    from typing import List
     from nltk.corpus import stopwords
     from nltk.stem.snowball import SnowballStemmer
     from nltk.tokenize import WhitespaceTokenizer
     import json
     import nltk
 
-    from model_bridging.helpers import (
-        get_bucket_and_key_from_s3_uri,
-        tokenize_pd_code,
-        get_date_diff,
-        add_emp_tenure_to_df,
-        download_model_from_s3,
-    )
-
     nltk.download("stopwords")
+
+    def get_bucket_and_key_from_s3_uri(uri):
+        bucket, key = uri.split("/", 2)[-1].split("/", 1)
+        return bucket, key
+
+    def tokenize_pd_code(df):
+
+        df[["pd_code_1", "pd_code_2"]] = pd.DataFrame(
+            df["PrimaryDiagnosisCode"].fillna("na.na").str.split(".", expand=True)
+        )
+        return df.drop("PrimaryDiagnosisCode", axis=1)
+
+    def get_date_diff(col1, col2, interval):
+        """difference between dates specified by the interval in ['Y','D','M']
+        col1 and col2 are date colummns and col2 > col1"""
+
+        return (col2 - col1) / np.timedelta64(1, interval)
+
+    def add_emp_tenure_to_df(df):
+        """
+        returns a df with employment tenure column appeneded
+        """
+        df["emp_tenure"] = get_date_diff(
+            df["InsuredHireDate"], df["LossDate"], interval="D"
+        )
+
+        return df.drop("InsuredHireDate", axis=1)
+
+
+    def download_model_from_s3(bucket_name, key):
+        bucket = boto3.resource("s3").Bucket(bucket_name)
+        with tempfile.NamedTemporaryFile() as fp:
+            bucket.download_fileobj(key, fp)
+            loaded_model = joblib.load(fp.name)
+        return loaded_model
 
     def stem_text(text):
         return " ".join([stemmer.stem(w) for w in w_tokenizer.tokenize(text)])
